requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

automaton>=1.9.0  # Apache-2.0
libvirt-python!=4.1.0,>=3.5.0 # LGPLv2+
openstacksdk>=0.13.0 # Apache-2.0
oslo.concurrency>=3.26.0 # Apache-2.0
oslo.config>=5.2.0 # Apache-2.0
lxml!=3.7.0,>=3.4.1  # BSD
oslo.cache>=1.26.0  # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.middleware>=3.31.0 # Apache-2.0
oslo.privsep>=1.23.0 # Apache-2.0
oslo.service!=1.28.1,>=1.24.0 # Apache-2.0
oslo.utils>=3.33.0 # Apache-2.0
pbr!=2.1.0,>=2.0.0 # Apache-2.0
python-openstackclient>=3.12.0 # Apache-2.0
six>=1.10.0 # MIT
